treure ngModels, form valids

+++ src/app/categoria/categoria.component.ts
@@ -1,6 +1,7 @@
 import { Component,  OnInit} from '@angular/core';
 import { Categoria } from 'src/app/modelo/categoria';
 import { CategoriaService } from '../service/categoria.service';
+import { SubcategoriaService } from '../service/subcategoria.service';
 import { Router } from '@angular/router'
 import { Dialogs } from 'src/app/dialogs/dialogs'
 import Swal from 'sweetalert2';
@@ -15,28 +16,25 @@ import Swal from 'sweetalert2';
 export class CategoriaComponent implements OnInit {
 
   categories!:Categoria[];
-  progress!:Boolean;
 
   displayedColumns: string[] = ['id', 'nom', 'descripcio','accions'];
   dataSource = this.categories;
 
-  constructor(private service:CategoriaService, private router:Router, private dialog:Dialogs){
-    this.progress = true;
+  constructor(private service:CategoriaService, private subCat_service: SubcategoriaService, private router:Router, private dialog:Dialogs){
   }
 
   ngOnInit(): void {
     this.service.getCategorias().subscribe
       (data=>{
         this.categories = data;
-    })   
-    this.progress = false;
+        this.categories.sort((x,y)=> x.id- y.id);
+      })   
   }
 
   Nou(){
     this.router.navigate(["editCat"]);
     this.service.isEdit = false;
   } 
-
 
   Editar(categoria:Categoria){
     this.router.navigate(["editCat/", categoria.id]);
@@ -54,12 +52,18 @@ export class CategoriaComponent implements OnInit {
       confirmButtonText: 'Si, borra\'l'
     }).then((result) => {
       if (result.isConfirmed) {
-            this.service.deleteCategoria(categoria).subscribe
-            (data=>{
-              this.categories= this.categories.filter(c=>c!==categoria);
-            })  
-
-        this.dialog.registregBorrat();    
+        this.subCat_service.ckSubCatToDeleteCategoria(categoria.id).subscribe
+        ((data)=>{ 
+          if (data == 0){
+              this.service.deleteCategoria(categoria).subscribe
+              (data=>{
+                this.categories= this.categories.filter(c=>c!==categoria);
+              })  
+            this.dialog.registregBorrat();  
+          }else{
+            this.dialog.simpleAlert("La categoria no es pot borrar ja que te "+ data +" subcategorias asociades","Categoria no borrable","info");
+          } 
+        })
       }
     })
   }
